% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_documents.R
\name{filter_documents}
\alias{filter_documents}
\title{Remove Documents Below a Threshold TermDocumentMatrix/DocumentTermMatrix}
\usage{
filter_documents(x, min = 1)
}
\arguments{
\item{x}{A \code{\link[tm]{TermDocumentMatrix}} or \code{\link[tm]{DocumentTermMatrix}}.}

\item{min}{A minimal threshold that a documents row/column must sum to.}
}
\value{
Returns a \code{\link[tm]{TermDocumentMatrix}} or \code{\link[tm]{DocumentTermMatrix}}.
}
\description{
Remove documents from a \code{\link[tm]{TermDocumentMatrix}}
or \code{\link[tm]{DocumentTermMatrix}} not meeting a \code{\link[base]{rowSums}}/
\code{\link[base]{colSums}} threshold.  Useful for removing empty documents.
}
\examples{
(x <-with(presidential_debates_2012, q_dtm(dialogue, paste(time, tot, sep = "_"))))
filter_documents(x)
(y <- with(presidential_debates_2012, q_tdm(dialogue, paste(time, tot, sep = "_"))))
filter_documents(y)
}

