% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gofastr-package.R
\docType{package}
\name{gofastr}
\alias{gofastr}
\alias{gofastr-package}
\alias{package-gofastr}
\title{Fast DocumentTermMatrix and TermDocumentMatric Creation}
\description{
This package does one thing...It harness the power of \pkg{data.table} and
\pkg{stringi} to quickly generate \pkg{tm} \code{\link[tm]{TermDocumentMatrix}}
and \code{\link[tm]{DocumentTermMatrix}} data structures without creating a
\code{\link[tm]{Corpus}} first.
}

