% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gofastr-package.R
\docType{data}
\name{partial_republican_debates_2015}
\alias{partial_republican_debates_2015}
\title{2015 U.S. Partial Republican Primary Presidential Debates}
\format{A data frame with 7405 rows and 5 variables}
\usage{
data(partial_republican_debates_2015)
}
\description{
A dataset containing a cleaned version of four primary presidential debates for
the 2016 election.
}
\details{
\itemize{
  \item location. Where debate took place
  \item person. The speaker
  \item dialogue. The words spoken
  \item element_id. Original line number (turn of talk) within location
  \item sentence_id. Sentence number within \code{element_id}
}
}
\references{
\url{http://www.presidency.ucsb.edu}
}
\keyword{datasets}
