% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/remove_stopwords.R
\name{remove_stopwords}
\alias{remove_stopwords}
\title{Remove Stopwords from a TermDocumentMatrix/DocumentTermMatrix}
\usage{
remove_stopwords(x, stopwords = tm::stopwords("english"), min.char = 3)
}
\arguments{
\item{x}{A \code{\link[tm]{TermDocumentMatrix}} or \code{\link[tm]{DocumentTermMatrix}}.}

\item{stopwords}{A vector of stopwords to remove.}

\item{min.char}{The minial length character for retained words.}
}
\value{
Returns a \code{\link[tm]{TermDocumentMatrix}} or \code{\link[tm]{DocumentTermMatrix}}.
}
\description{
Remove stopwords and < nchar words from a \code{\link[tm]{TermDocumentMatrix}}
or \code{\link[tm]{DocumentTermMatrix}}.
}
\examples{
(x <-with(presidential_debates_2012, q_dtm(dialogue, paste(time, tot, sep = "_"))))
remove_stopwords(x)
(y <- with(presidential_debates_2012, q_tdm(dialogue, paste(time, tot, sep = "_"))))
remove_stopwords(y)
}
\keyword{stopwords}

