% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_dtm.R
\name{as_dtm}
\alias{as_dtm}
\alias{as_tdm}
\title{Coerce Various Object Into a \code{DocumentTermMatrix}/\code{TermDocumentMatrix}}
\usage{
as_dtm(x, weighting = tm::weightTf, ...)

as_tdm(x, weighting = tm::weightTf, ...)
}
\arguments{
\item{x}{A data object.}

\item{weighting}{A weighting function capable of handling a \code{tm::DocumentTermMatrix}. It defaults to \code{weightTf} for term frequency weighting. Available weighting functions shipped with the \pkg{tm} package are \code{weightTf}, \code{weightTfIdf}, \code{weightBin}, and \code{weightSMART}.}

\item{\ldots}{ignored.}
}
\value{
Returns a \code{tm::DocumentTermMatrix} or \code{tm::TermDocumentMatrix} object.
}
\description{
Convenience functions to convert a objects from different packages into either a \code{tm::DocumentTermMatrix} or \code{tm::TermDocumentMatrix} object.  Grouping variables are used as the row/column names for the \code{DocumentTermMatrix}/\code{TermDocumentMatrix}.
}
\examples{
x <- 'temp holder'
}
\keyword{documenttermmatrix,}
\keyword{termdocumentmatrix}

