% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/q_dtm.R
\name{q_dtm}
\alias{q_dtm}
\title{Quick DocumentTermMatrix}
\usage{
q_dtm(text, docs = seq_along(text), weighting = tm::weightTf, ...)
}
\arguments{
\item{text}{A vector of strings.}

\item{docs}{An optional vector of document names.}

\item{weighting}{A \pkg{tm} weighting: \code{\link[tm]{weightTf}},
\code{\link[tm]{weightTfIdf}}, \code{\link[tm]{weightBin}}, or
\code{\link[tm]{weightSMART}}.}

\item{\ldots}{Additional arguments passed to \code{\link[tm]{as.DocumentTermMatrix}}.}
}
\value{
Returns a \code{\link[tm]{DocumentTermMatrix}}.
}
\description{
Make a \code{\link[tm]{DocumentTermMatrix}} from a vector of text and and
optional vector of documents.
}
\examples{
(x <- with(presidential_debates_2012, q_dtm(dialogue, paste(time, tot, sep = "_"))))
tm::weightTfIdf(x)
}
\keyword{DocumentTermMatrix}
\keyword{dtm}

