% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/q_tdm.R
\name{q_tdm}
\alias{q_tdm}
\title{Quick TermDocumentMatrix}
\usage{
q_tdm(text, docs = seq_along(text), weighting = tm::weightTf, ...)
}
\arguments{
\item{text}{A vector of strings.}

\item{docs}{An optional vector of document names.}

\item{weighting}{A \pkg{tm} weighting: \code{\link[tm]{weightTf}},
\code{\link[tm]{weightTfIdf}}, \code{\link[tm]{weightBin}}, or
\code{\link[tm]{weightSMART}}.}

\item{\ldots}{Additional arguments passed to \code{\link[tm]{as.TermDocumentMatrix}}.}
}
\value{
Returns a \code{\link[tm]{TermDocumentMatrix}}.
}
\description{
Make a \code{\link[tm]{TermDocumentMatrix}} from a vector of text and and
optional vector of documents.
}
\examples{
(x <- with(presidential_debates_2012, q_tdm(dialogue, paste(time, tot, sep = "_"))))
tm::weightTfIdf(x)
}
\keyword{TermDocumentMatrix}
\keyword{tdm}

